% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benito_paula}
\alias{benito_paula}
\title{benito_paula}
\format{A data frame with 6 variables:
\describe{
  \item{\code{date}}{integer.The date of the album which contains the
  music.}
  \item{\code{music}}{factor. The name of the music. }
  \item{\code{popul}}{integer. The popularity of the music.}
  \item{\code{chord}}{factor. The chord names of each music,
  by order of occurrence in the music. }
  \item{\code{key}}{factor. The key for each music.}
  \item{\code{artist}}{factor. The name of the artist}
}}
\usage{
benito_paula
}
\description{
Data with information about the music chords of Benito di Paula.
}
\keyword{datasets}
