% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bruno_marrone}
\alias{bruno_marrone}
\title{bruno_marrone
Data with information about the music chords of Bruno & Marrone.}
\format{A data frame with 6 variables:
\describe{
  \item{\code{date}}{integer.The date of the album which contains the
  music.}
  \item{\code{music}}{factor. The name of the music. }
  \item{\code{popul}}{integer. The popularity of the music.}
  \item{\code{chord}}{factor. The chord names of each music,
  by order of occurrence in the music. }
  \item{\code{key}}{factor. The key for each music.}
  \item{\code{artist}}{factor. The name of the artist}
}}
\usage{
bruno_marrone
}
\description{
bruno_marrone
Data with information about the music chords of Bruno & Marrone.
}
\keyword{datasets}
