% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords_bigram.R
\name{chords_ngram}
\alias{chords_ngram}
\title{chords_ngram}
\usage{
chords_ngram(data, n = 2)
}
\arguments{
\item{data}{dataframe. The chords dataset to exract the features
from.}

\item{n}{numeric. The number of grams. The default is 2 
(bigram).}
}
\value{
A chords dataset added with the chords ngram.
}
\description{
Builds chords ngrams for a chord datdaset.
}
\examples{
{
\donttest{
  songs <- chorrrds::get_songs("tim-maia")
  chords <- get_chords(songs$url[4])
  chords_ngram(chords)
}
}
}
