% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dat.R
\name{create_dat}
\alias{create_dat}
\title{create_dat}
\usage{
create_dat(artist, track)
}
\arguments{
\item{artist}{character. The artist's name.}

\item{track}{character. The song's title.}
}
\value{
An object of type `data.frame` with the song chords and lyrics is retuned. 
The object is to be later used in the `create_net()` function to get accurate connections between chords and words.
}
\description{
Break song by verse with chords and corresponding lyrics.
}
\examples{
{
\donttest{
  create_dat("The Weeknd", "Acquainted")
}
}
}
