% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_net.R
\name{create_net}
\alias{create_net}
\title{create_net}
\usage{
create_net(chords_dat)
}
\arguments{
\item{chords_dat}{data frame. A data frame as produced by the `create_dat()` function with chords in the first column and lyrics in the second column.}
}
\value{
An object of type `tibble` with the song chords and lyrics is retuned.
Each chord is linked to the words that are sung when that chord is played.
}
\description{
Match music lyrics with the corresponding chords.
}
\examples{
{
\donttest{
  chords_dat <- create_dat("The Weeknd", "Acquainted")
  create_net(chords_dat)
}
}
}
