% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_min}
\alias{deg_min}
\title{deg_min}
\format{An object of class \code{data.frame} with 7 rows and 16 columns.}
\usage{
deg_min
}
\description{
Accessory data with the chords present in each scale, with
its respective degrees, for the minor cases.
}
\keyword{datasets}
