% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist}
\alias{dist}
\title{dist}
\format{A data frame with 3 variables:
\describe{
  \item{\code{prox}}{factor. The chord.}
  \item{\code{dist}}{numeric. The distance from C in the
  circle of fifths. }
  \item{\code{order}}{integer. The order in the circle of
  fifths.}
}}
\usage{
dist
}
\description{
A simple measure of the chords distances in the circle
of fifths.
}
\keyword{datasets}
