% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eqv}
\alias{eqv}
\title{eqv}
\format{A data frame with 3 variables:
\describe{
  \item{\code{key}}{factor. Keys ordered by the circle of fifths.}
  \item{\code{minor.rel}}{factor. Relative minors of the key
  in the previous column. }
  \item{\code{rep}}{num. A number indicating if the key scale is
  equivalent to some other; repeated numbers indicate equivalent
  keys. }
}}
\usage{
eqv
}
\description{
Accessory data for the recognition of equivalent keys, including major
and minor relatives.
}
\keyword{datasets}
