% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{feature_extraction}
\alias{feature_extraction}
\title{feature_extraction}
\usage{
feature_extraction(data)
}
\arguments{
\item{data}{dataframe. The chords dataset to exract the features
from.}
}
\value{
A dataframe with the chords set add with  
logical features (1 or 0), to indicate if each chord is:
}
\description{
Extracts features from a chords dataset.
}
\examples{
{
\donttest{
  songs <- get_songs("tim-maia")
  chords <- get_chords(songs$url[4])
  feature_extraction(chords)
}
}
}
