% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genre}
\alias{genre}
\title{genre}
\format{An object of class \code{data.frame} with 106 rows and 2 columns.}
\usage{
genre
}
\description{
Accessory data with the genre for each artist in the package.
}
\keyword{datasets}
