% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs.R
\name{get_songs}
\alias{get_songs}
\title{get_songs}
\usage{
get_songs(artist)
}
\arguments{
\item{artist}{character. The artist's name.}
}
\value{
If the artist (or band) is found, an object of type 
`tibble` with the song names, URLs  and artist is retuned. 
The URLs are to be later used in the `get_chords()` function.
}
\description{
Get songs names and URLs for an artist.
}
\examples{
{
\donttest{
  get_songs("jorge")
  get_songs("los-hermanos")
}
}
}
