% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_in}
\alias{not_in}
\title{This function negates the macth operator.}
\usage{
not_in(x, y)
}
\description{
This function negates the macth operator.
}
\examples{
{
\dontrun{
  not_in(this_is_not, in_this)
 }
}
}
\keyword{internal}
