% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\alias{.}
\title{Pipe operator}
\description{
See \code{\link[magrittr]{\%>\%}} for more details.

See \code{\link[plyr]{.}} for more details.
}
\examples{
{

  iris \%>\% as.matrix()
}
}
\keyword{internal}
