% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{search_data}
\usage{
search_data(name)
}
\arguments{
\item{name}{character. The searched artist's name.}
}
\value{
If a database with the corresponding searched name
        is found, it's name is returned. If not, nothing is returned.
}
\description{
Search artists in the available package database.
}
\examples{
{
   search_data("chico")
  \dontshow{
  search_data("jorge")
  }
}
}
