% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{'\%!in\%'}
\alias{'\%!in\%'}
\alias{\%!in\%}
\title{This function negates the %in% operator.}
\usage{
this_is_not '%!in%' in_this
}
\description{
This function negates the %in% operator.
}
\keyword{internal}
