% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(da, column = "chord", long = 15, message = TRUE)
}
\arguments{
\item{da}{a data.frame.}

\item{column}{string. The column by which we want to make the cleaning.}

\item{long}{numeric. The longest string we wish exists on our}

\item{message}{logical. Should the function print how many lines
were removed?}
}
\value{
A database, with the text cleaning done.
}
\description{
Clean data when there is some excessive long text on a column.
}
\examples{
{
\dontrun{
data("caetano")
  clean(da  = caetano, column = "chord", long = 15, message = TRUE)
}
}
}
