% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chords.R
\name{get_chords}
\alias{get_chords}
\title{get_chords}
\usage{
get_chords(song_url, nf = FALSE)
}
\arguments{
\item{song_url}{character. The song url.}

\item{nf}{TRUE of FALSE. If the chords of a song are not found,
should we return this information on the database?}
}
\value{
A database with the chords of the input songs, if
found.
}
\description{
Extracts music chords from an artist.
}
\examples{
{
\donttest{
  songs <- get_songs("tim-maia")
  get_chords(songs$url[2])
}
}
}
