% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs.R
\name{get_songs}
\alias{get_songs}
\title{get_songs}
\usage{
get_songs(artist)
}
\arguments{
\item{artist}{character. The artist's name.}
}
\value{
If the artist (or band) is found, it is returned a database
with the song names and URLs.
}
\description{
Get songs names from an artist.
}
\examples{
{
\donttest{
  get_songs("jorge")
  get_songs("los-hermanos")
}
}
}
