% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_in}
\alias{not_in}
\title{This function negates the match operator.}
\usage{
not_in(x, y)
}
\description{
This function negates the match operator.
}
\examples{
{
a <- "a"
ab <- c("a", "b")
  not_in(a, ab)
}
}
\keyword{internal}
