% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{'\%!in\%'}
\alias{'\%!in\%'}
\alias{\%!in\%}
\title{This function negates the %in% operator.}
\usage{
x \%!in\% y
}
\description{
This function negates the %in% operator.
}
\examples{
{
\dontrun{
  this_is_not '\%!in\%' in_this
 }
}
}
\keyword{internal}
